pyproject.toml
[project]
name = "litdata-with-minio"
version = "0.1.0"
description = "A demonstration of using Lightning AI's LitData with a MinIO S3-compatible object store."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Bhimraj Yadav", email = "bhimrajyadav977@gmail.com" }]
dependencies = ["awscli", "litdata", "numpy", "pillow", "python-dotenv"]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
